refactor: move the metadata auth token to core
